refactor: share helper session and clean tests
